Moving and reordering some tests around browser client.
