+++ src/__tests__/browserSuites/events.spec.js
@@ -1,10 +1,10 @@
 'use strict';
 
-const SplitFactory = require('../');
+const SplitFactory = require('../../');
 
 const fetchMock = require('fetch-mock');
 
-const SettingsFactory = require('../utils/settings');
+const SettingsFactory = require('../../utils/settings');
 const settings = SettingsFactory({
   core: {
     key: 'asd'
